app/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.44
scikit-learn==1.5.2
pandas-ta==0.3.14b0
SQLModel==0.0.22
aiosqlite==0.20.0
nltk==3.9.1
requests==2.32.3
python-dotenv==1.0.1
